Refactor project index root discovery
